cymbal_agent/requirements.txt
# --- Core Vertex/ADK client libs (ADK's Dockerfile already installs google-adk==1.16.0) ---
google-cloud-aiplatform[agent_engines,adk,evaluation]>=1.124.0
google-genai>=1.47.0
google-cloud-storage>=2.19.0


# --- Parsing / utils ---
beautifulsoup4>=4.12.0
PyPDF2>=3.0.0
requests>=2.32.0
python-dotenv>=1.0.0

# --- Database (Cloud SQL MySQL) ---
SQLAlchemy>=2.0.0
PyMySQL>=1.1.0
cloud-sql-python-connector[pymysql]>=1.9.0

# -- Other ---
# google-cloud-vision
# Pillow
